Adaptations de la création de bien

+++ app/user/user.controller.js
@@ -9,8 +9,9 @@ const User = require('./users.model');
 const jwt = require('../helpers/jwt');
 
 
-findById = (req, res) => {
+userFindById = (req, res) => {
     userService.findById(req.params.id, (err, data) => {
+        console.log(req.params.id);
         if (err) {
             if (err.kind === "not_found") {
                 res.status(404).send({
@@ -26,47 +27,55 @@ findById = (req, res) => {
 };
 
 post = (req, res) => {
+    console.log('jeminscrit');
     var user = new User(req.body);
     console.log(user);
-    bcrypt.hash(user.Password, 5).then(function (hash) {
-        user.Password = hash;
+    bcrypt.hash(user.MDP, 5).then(function (hash) {
+        user.MDP = hash;
         userService.signup(user, (err, data) => {
             if (err) {
                 if (err.kind === "not_found") {
-                    res.status(200).send({
-                        message: `Not found user with id ${req.params.customerId}.`
+                    res.status(204).send({
+                        message: `utilisateur créée.`
                     });
                 } else {
                     res.status(500).send({
-                        message: "Error retrieving user with id " + req.params.customerId
+                        message: "Error retrieving user creation"
                     });
                 }
             } else res.send(data);
         });
-
     });
 };
 
 put = (req, res) => {
-    userService.put(req.params.id, req.body.Nom, req.body.Prenom, req.body.Mail, req.body.Telephone, (err, data) => {
-        if (err) {
-            if (err.kind === "not_found") {
-                res.status(404).send({
-                    message: `Not found user with id ${req.params.customerId}.`
-                });
+    if(parseInt(req.params.id) == jwt.getUserId(req.headers.authorization)) {
+        console.log("je rentre");
+        userService.put(req.params.id, req.body.Nom, req.body.Prenom, req.body.Mail, req.body.Telephone, (err, data) => {
+            if (err) {
+                if (err.kind === "not_found") {
+                    res.status(404).send({
+                        message: `Not found user with id ${req.params.id}.`
+                    });
+                } else {
+                    res.status(500).send({
+                        message: "Error retrieving user with id " + req.params.id
+                    });
+                }
             } else {
-                res.status(500).send({
-                    message: "Error retrieving user with id " + req.params.customerId
-                });
+                return userFindById(req,res);
             }
-        } else res.send(data);
-    });
+        });
+    } else {
+        return res.status(403).send({
+            message: `Vous ne pouvez pas modifier l'utilisateur`
+        });
+    }
 };
 
 login = (req, res) => {
-    var mail = req.body.mail;
-    var password = req.body.password;
-    // check if userfound
+    var mail = req.body.Mail;
+    var password = req.body.MDP;
     userService.findByEmail(mail, (err, userFound) => {
         if (userFound) {
             bcrypt.compare(password, userFound.MDP, function (errBycript, resBycript) {
@@ -76,33 +85,39 @@ login = (req, res) => {
                         'token': jwt.generateTokenForUser(userFound),
                     })
                 } else {
-                    return res.status(403).json({'error': 'invalid password'});
+                    return res.status(403).json({message: 'Information de connexion incorrectes'});
                 }
             })
         } else {
-            return res.status(403).json({'error': 'user not found'});
+            return res.status(403).json({message: "Information de connexion incorrectes"});
         }
     })
 }
 
 deleteUser = (req, res) => {
-    userService.deleteUser(req.params.id, (err, data) => {
-        if (err) {
-            if (err.kind === "not_found") {
-                res.status(404).send({
-                    message: `Not found user with id ${req.params.customerId}.`
-                });
-            } else {
-                res.status(500).send({
-                    message: "Error retrieving user with id " + req.params.customerId
-                });
-            }
-        } else res.send(data);
-    });
+    if(parseInt(req.params.id) == jwt.getUserId(req.headers.authorization)) {
+        userService.deleteUser(req.params.id, (err, data) => {
+            if (err) {
+                if (err.kind === "not_found") {
+                    res.status(404).send({
+                        message: `Not found user with id ${req.params.customerId}.`
+                    });
+                } else {
+                    res.status(500).send({
+                        message: "Error retrieving user with id " + req.params.customerId
+                    });
+                }
+            } else res.send(data);
+        });
+    } else {
+        return res.status(403).send({
+            message: `Vous ne pouvez pas supprimer l'utilisateur`
+        });
+    }
 };
 
 
-router.get('/:id', findById);
+router.get('/:id', userFindById);
 router.post('/login/', login);
 
 router.post('/', post);
